API/Reference

PSFKernels.PSFKernels

PSFKernels.AiryDisk

PSFKernels.Gaussian

PSFKernels.Moffat

PSFKernels.Normal

PSFKernels.PSFKernel

PSFKernels.PSFKernel — Type

Kernels.PSFKernel{T} <: AbstractMatrix{T}

Abstract type for PSF Kernels.

In general, all PSFKernels have a set of pre-determined axes (the size is set upon creation) but they are lazy. That is, no memory is allocated and the values are calculated on the fly.

Interface

The interface to define a kernel is as follows (for an example kernel Kernel)

method	description

Kernel()	constructor(s)

Base.size(k::Kernel)	size, necessary for AbstractArray interface

Base.axes(k::Kernel)	axes, necessary for AbstractArray interface

(k::Kernel)(point::AbstractVector)	evaluate the kernel at the point in 2d space

browsing through the implementation of PSFKernels.Gaussian should give a good idea of how to create a kernel

Gaussian

PSFKernels.Gaussian — Type

PSFKernels.Gaussian(fwhm; maxsize=3)
PSFKernels.Gaussian(position, fwhm; maxsize=3)
PSFKernels.Gaussian(x, y, fwhm; maxsize=3)
PSFKernels.Gaussian(::Polar, fwhm; maxsize=3, origin=(0, 0))
PSFKernels.Gaussian{T}(args...; kwargs...)

An unnormalized bivariate Gaussian distribution. The position can be specified in (x, y) coordinates as a Tuple, AbstractVector, or as separate arguments. By default the kernel is placed at the origin. The position can also be given as a CoordinateTransformations.Polar, optionally centered around origin.

The fwhm can be a scalar (isotropic), vector/tuple (diagonal), or a matrix (correlated). For efficient calculations, we recommend using StaticArrys. Here, maxsize is a multiple of the fwhm, and can be given as a scalar or as a tuple for each axis.

The output type can be specified, and will default to Float64. The amplitude is unnormalized, meaning the maximum value will always be 1. This is distinct from the probability distribution (pdf) of a bivariate Gaussian which assures the kernel sums to 1. This means the kernels act like a transmission weighting instead of a probability weighting.

Functional form

f(x | x̂, FWHM) = exp[-4ln(2) * ||x - x̂|| / FWHM^2]

where x̂ and x are position vectors (indices) ||⋅|| represents the square-distance, and FWHM is the full width at half-maximum. If FWHM is a vector or tuple, the weighting is applied along each axis.

If the FWHM is a correlated matrix, the functional form uses the square-Mahalanobis distance

f(x | x̂, Q) = exp[-4ln(2) * (x - x̂)ᵀ Q (x - x̂)]

where Q is the inverse covariance matrix (or precision matrix). This is equivalent to the inverse of the FWHM matrix after squaring each element.

PSFKernels.Normal — Type

PSFKernels.Normal

An alias for PSFKernels.Gaussian

kernel = Gaussian(10)
plot(kernel; title="Gaussian(fwhm=10)")

Airy Disk

PSFKernels.AiryDisk — Type

PSFKernels.AiryDisk(fwhm; maxsize=3)
PSFKernels.AiryDisk(position, fwhm; maxsize=3)
PSFKernels.AiryDisk(x, y, fwhm; maxsize=3)
PSFKernels.AiryDisk(::Polar, fwhm; maxsize=3, origin=(0, 0))
PSFKernels.AiryDisk{T}(args...; kwargs...)

An unnormalized Airy disk. The position can be specified in (x, y) coordinates as a Tuple, AbstractVector, or as separate arguments. By default the kernel is placed at the origin. The position can also be given as a CoordinateTransformations.Polar, optionally centered around origin.

The fwhm can be a scalar (isotropic) or vector/tuple (diagonal). For efficient calculations, we recommend using StaticArrys. Here, maxsize is a multiple of the fwhm, and can be given as a scalar or as a tuple for each axis.

The output type can be specified, and will default to Float64. The amplitude is unnormalized, meaning the maximum value will always be 1. This means the kernels act like a transmission weighting.

Functional form

TODO

kernel = AiryDisk(10)
plot(kernel; title="AiryDisk(fwhm=10)")

Moffat

PSFKernels.Moffat — Type

PSFKernels.Moffat(fwhm; maxsize=3)
PSFKernels.Moffat(position, fwhm; maxsize=3)
PSFKernels.Moffat(x, y, fwhm; maxsize=3)
PSFKernels.Moffat(::Polar, fwhm; maxsize=3, origin=(0, 0))
PSFKernels.Moffat{T}(args...; kwargs...)

An unnormalized Airy disk. The position can be specified in (x, y) coordinates as a Tuple, AbstractVector, or as separate arguments. By default the kernel is placed at the origin. The position can also be given as a CoordinateTransformations.Polar, optionally centered around origin.

The fwhm can be a scalar (isotropic) or vector/tuple (diagonal). For efficient calculations, we recommend using StaticArrys. Here, maxsize is a multiple of the fwhm, and can be given as a scalar or as a tuple for each axis.

The output type can be specified, and will default to Float64. The amplitude is unnormalized, meaning the maximum value will always be 1. This means the kernels act like a transmission weighting.

Functional form

f(x | x̂, FWHM) = 1 / [1 + ||x - x̂|| / (FWHM / 2)^2]

where x̂ and x are position vectors (indices) ||⋅|| represents the square-distance, and FWHM is the full width at half-maximum. If FWHM is a vector or tuple, the weighting is applied along each axis.

kernel = Moffat(10)
plot(kernel; title="Moffat(fwhm=10)")