PSFKernels.jl

Installation

PSFKernels can be added from the Julia package manager

julia>]

(@v1.6) pkg> add PSFKernels

Getting Started

To import the library

julia> using PSFKernels

None of the kernels are exported to avoid namespace clashes, but it can be verbose. You can either import names directly

julia> using PSFKernels: Gaussian

julia> kernel = Gaussian(8)

or you can create an alias for PSFKernels

# julia version 1.5 or below
using PSFKernels
const kerns = PSFKernels
# julia version 1.6 or above
using PSFKernels as kerns

kernel = kerns.Gaussian(10)

PSFKernels.PSFKernels — Module

PSFKernels

Statistical kernels for constructing point-spread functions (PSFs). These kernels act like matrices but without allocating any memory, which makes them efficient to fit and apply.

Kernels

The following kernels are currently implemented

PSFKernels.Gaussian

PSFKernels.AiryDisk

PSFKernels.Moffat

Usage

Using the kernels should feel just like an array. In fact, PSFKernels.PSFKernel <: AbstractMatrix. However, no data is stored and no allocations have to be made. In other words, representing the kernels as matrices is merely a convenience, since typically astronomical data is stored in dense arrays.

julia> k = PSFKernels.Gaussian(5); # fwhm of 5 pixels, centered at (0, 0)

julia> k[0, 0]
1.0

because the kernel is a matrix, it needs to have a size. In this case, the size is maxsize * FWHM pixels, centered around the origin, and rounded up. We can see how this alters the indices from a typical Matrix

julia> size(k)
(17, 17)

julia> axes(k)
(-8:8, -8:8)

if we want to collect the kernel into a dense matrix, regardless of the indexing (e.g. to prepare for cross-correlation), we can simply

julia> stamp = collect(k);

these axes are merely a convenience for bounding the kernel, since they accept any real number as input.

julia> k[100, 10000] # valid for index-like inputs
0.0

julia> k(2.4, 1.7) # valid for any number
0.38315499005194587

By bounding the kernel, we get a cutout which can be applied to arrays with much larger dimensions without having to iterate over the whole matrix

Coordinate System

The PSFs follows a 1-based image coordinate system, where (1, 1) represents the center of the bottom left pixel. This corresponds with Julia indexing, as well as DS9 and IRAF.

julia> big_mat = ones(101, 101);

julia> small_kern = PSFKernels.Gaussian(51, 51, 2); # center of big_mat, fwhm=2

julia> ax = map(intersect, axes(big_mat), axes(small_kern))
(48:54, 48:54)

julia> cutout = big_mat[ax...]
7×7 Array{Float64,2}:
 1.0  1.0  1.0  1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0  1.0  1.0  1.0

julia> photsum = sum(cutout .* small_kern[ax...])
4.5322418212890625

Nice- we only had to reduce ~50 pixels instead of ~10,000 to calculate the aperture sum, and with some care we could make it allocation-free.

Since the kernels are lazy, that means the type of the output can be specified, as long as it can be converted to from a real number (so no integer types).

julia> kbig = PSFKernels.Gaussian{BigFloat}(12);

julia> sum(kbig)
163.07467408408593790971336380361822460116627553361468017101287841796875

finally, we provide plotting recipes from RecipesBase.jl, which can be seen in use in the API/Reference section.

using Plots
kernel = PSFKernels.Gaussian(8)
plot(kernel)              # default axes
plot(kernel, 1:5, 1:5)    # custom axes
plot(kernel, axes(other)) # use axes from other array

Tip: Automatic Differentation

Forward-mode AD libraries tend to use dual numbers, which can cause headaches getting the types correct. We recommend using the primal vector's element type to avoid these headaches

# example generative model for position and scalar fwhm
kernel(X::AbstractVector{T}) where {T} = PSFKernels.Gaussian{T}(X...)